refactor: split playback completion handlers
