+++ app/shared/player/hooks/singleRepeatHandler.ts
@@ -1,0 +1,38 @@
+interface SingleRepeatArgs {
+  verseRepeatsLeft: number;
+  playRepeatsLeft: number;
+  repeatEach: number;
+  seek: (s: number) => void;
+  play: () => void;
+  setVerseRepeatsLeft: (n: number) => void;
+  setPlayRepeatsLeft: (n: number) => void;
+}
+
+export function createSingleRepeatHandler({
+  verseRepeatsLeft,
+  playRepeatsLeft,
+  repeatEach,
+  seek,
+  play,
+  setVerseRepeatsLeft,
+  setPlayRepeatsLeft,
+}: SingleRepeatArgs): () => boolean {
+  return () => {
+    if (verseRepeatsLeft > 1) {
+      setVerseRepeatsLeft(verseRepeatsLeft - 1);
+      seek(0);
+      play();
+      return true;
+    }
+    if (playRepeatsLeft > 1) {
+      setPlayRepeatsLeft(playRepeatsLeft - 1);
+      setVerseRepeatsLeft(repeatEach);
+      seek(0);
+      play();
+      return true;
+    }
+    return false;
+  };
+}
+
+export type { SingleRepeatArgs };
